pyproject.toml
[project]
name = "multi-pdf-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.0.0",
    "docling>=2.68.0",
    "faiss-cpu>=1.8.0",
    "huggingface-hub>=0.23.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-docling>=2.0.0",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=0.1.5",
    "langchain-huggingface>=1.2.0",
    "mysql>=0.0.3",
    "pandas>=2.3.3",
    "pymysql>=1.1.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=2.6.0",
    "sqlglot>=28.6.0",
    "streamlit>=1.32.0",
]
